v0.2.0: 文献管理模块、PDF图片转换、题库学习优化、云同步修复
新增:
- 文献管理（PDF导入/在线搜索/DOI导入/标注笔记/引用格式/AI摘要）
- PDF ↔ 图片转换
- 题库筛选后直接学习 + 添加题目默认章节题型
- 媒体压缩支持小数目标大小

修复:
- 多账号数据隔离（wechat_open_id优先匹配）
- 云同步删除传播（防止已删除内容被同步回来）
- 模拟买入净值为0的问题
- DOI导入格式兼容
- 移除水印功能

+++ app/src/main/java/com/psymap/app/DocumentProcessor.kt
@@ -774,7 +774,7 @@ object DocumentProcessor {
         }
 
         // 用二分法找到最接近目标大小的质量
-        val targetBytes = params.targetSizeMB * 1024L * 1024L
+        val targetBytes = (params.targetSizeMB * 1024 * 1024).toLong()
         var low = 10
         var high = 95
         var bestBytes: ByteArray? = null
@@ -833,7 +833,7 @@ object DocumentProcessor {
 
             // 计算目标比特率（根据目标文件大小）
             val targetBitrate = if (params.targetSizeMB > 0 && durationMs > 0) {
-                val targetBits = params.targetSizeMB.toLong() * 1024 * 1024 * 8
+                val targetBits = (params.targetSizeMB * 1024 * 1024 * 8).toLong()
                 val durationSec = durationMs / 1000.0
                 // 预留10%给音频
                 ((targetBits * 0.9) / durationSec).toInt().coerceIn(500_000, 50_000_000)
